fix main toolbars

+++ Dropin/04_Domain/UseCases/DeletePlace.swift
@@ -16,6 +16,9 @@ struct DeletePlace {
     }
     
     func execute(_ place: PlaceEntity) async throws {
+        if try await !repository.exists(place) {
+            throw DomainError.Place.notFound
+        }
         return try await repository.delete(place)
     }
 }
